add confirm dialog

+++ src/Main/java/Component/SettingComponent/Bookmark.java
@@ -5,7 +5,11 @@ import DAO.BookmarkDAO;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
 import javafx.collections.ObservableList;
+import javafx.geometry.Pos;
+import javafx.scene.control.Alert;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
+import javafx.scene.layout.HBox;
 
 // Data class for bookmark table
 public class Bookmark {
@@ -13,7 +17,7 @@ public class Bookmark {
     private StringProperty name;
     private StringProperty url;
     private StringProperty timeAdded;
-    private Button deleteBtn;
+    private HBox buttonGroup;
     private ObservableList<Bookmark> items;
 
     public Bookmark(String id, String name, String url, String timeAdded, ObservableList<Bookmark> items) {
@@ -22,15 +26,42 @@ public class Bookmark {
         this.url = new SimpleStringProperty(url);
         this.timeAdded = new SimpleStringProperty(timeAdded);
         this.items = items;
-        this.deleteBtn = new Button();
-        this.deleteBtn.getStyleClass().add("delete");
-        this.deleteBtn.getStylesheets().add("css/Style.css");
-        this.deleteBtn.setOnAction((actionEvent -> {
-            BookmarkDAO bookmarkDAO = new BookmarkDAO(SessionManager.getInstance().getUserId());
-            bookmarkDAO.deleteBookmark(id);
-            SessionManager.getInstance().updateBookmarkList();
-            items.remove(this);
+        this.buttonGroup = new HBox();
+
+        Button deleteBtn = new Button();
+        deleteBtn.getStyleClass().add("delete");
+        deleteBtn.getStylesheets().add("css/Style.css");
+        deleteBtn.setOnAction((actionEvent -> {
+            Alert confirm = new Alert(Alert.AlertType.CONFIRMATION, "Are you sure you want to delete this bookmark?");
+            confirm.showAndWait().ifPresent((response) -> {
+                if(response == ButtonType.OK) {
+                    BookmarkDAO bookmarkDAO = new BookmarkDAO(SessionManager.getInstance().getUserId());
+                    bookmarkDAO.deleteBookmark(id);
+                    SessionManager.getInstance().updateBookmarkList();
+                    items.remove(this);
+                }
+            });
         }));
+
+        Button editButton = new Button();
+        editButton.getStyleClass().add("edit");
+        editButton.getStyleClass().add("css/Style.css");
+        editButton.setOnAction((actionEvent -> {
+            for(int index = 0; index < items.size(); index++) {
+                Bookmark curr = items.get(index);
+                if(curr == this){
+                    String newName = new BookmarkDialog(curr.getId(),curr.getName()).showAndWaitForResponse();
+                    if(newName != null) {
+                        curr.setName(newName);
+                        items.set(index, curr);
+                    }
+                    break;
+                }
+            }
+        }));
+        this.buttonGroup.getChildren().addAll(deleteBtn, editButton);
+        this.buttonGroup.setAlignment(Pos.CENTER);
+        this.buttonGroup.setSpacing(10);
 
     }
 
@@ -50,8 +81,8 @@ public class Bookmark {
         return timeAdded.get();
     }
 
-    public Button getDeleteBtn() {
-        return deleteBtn;
+    public HBox getButtonGroup() {
+        return buttonGroup;
     }
 
     public void setName(String name) {
